FIX type des nombres d'entrées : unsigned int de 64 bits pour les nombres venant de l'entrée et unsigned int de 32 bits pour les nombres premiers

+++ src/factorisation.h
@@ -9,6 +9,8 @@
 #ifndef FACTORISATION_FACTORISATION_H
 #define FACTORISATION_FACTORISATION_H
 
+#include <stdint.h>
+
 /*
  * structure représentant un facteur premier
  *
@@ -18,7 +20,7 @@
  */
 struct facteurPremier
 {
-	unsigned int nombre;
+	uint32_t nombre;
 	int multiplicite;
 	char ** file;
 };
@@ -42,7 +44,7 @@ struct listeFacteurPremier
  */
 struct nombre
 {
-	unsigned int nombre;
+	uint64_t nombre;
 	char * file;
 };
